requirements.txt
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.6
python-decouple==3.3
whitenoise==5.2.0
